Caching the NBP API results
