+++ src/main/java/pl/cleankod/exchange/provider/CurrencyConversionNbpService.java
@@ -1,5 +1,6 @@
 package pl.cleankod.exchange.provider;
 
+import org.springframework.cache.annotation.Cacheable;
 import pl.cleankod.exchange.core.domain.Money;
 import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
 import pl.cleankod.exchange.provider.nbp.ExchangeRatesNbpClient;
@@ -17,6 +18,7 @@ public class CurrencyConversionNbpService implements CurrencyConversionService {
     }
 
     @Override
+    @Cacheable("money")
     public Money convert(Money money, Currency targetCurrency) {
         RateWrapper rateWrapper = exchangeRatesNbpClient.fetch("A", targetCurrency.getCurrencyCode());
         BigDecimal midRate = rateWrapper.rates().get(0).mid();
